+++ src/api/cpp/cvc5.cpp
@@ -63,6 +63,7 @@
 #include "options/options_public.h"
 #include "options/quantifiers_options.h"
 #include "options/smt_options.h"
+#include "proof/proof_node.h"
 #include "proof/unsat_core.h"
 #include "smt/env.h"
 #include "smt/model.h"
@@ -5016,6 +5017,76 @@ std::ostream& operator<<(std::ostream& out, const Statistics& stats)
   return out;
 }
 
+/*--------------------------------------------------------------------------- */
+/* Proof                                                                      */
+/* -------------------------------------------------------------------------- */
+
+Proof::Proof() {}
+
+Proof::Proof(const std::shared_ptr<internal::ProofNode> p) : d_proof_node(p) {}
+
+Proof::~Proof()
+{
+  Assert(d_proof_node != nullptr);
+  d_proof_node.reset();
+}
+
+ProofRule Proof::getRule() const
+{
+  CVC5_API_TRY_CATCH_BEGIN;
+  //////// all checks before this line
+  return this->getProofNode()->getRule();
+  ////////
+  CVC5_API_TRY_CATCH_END;
+}
+
+Term Proof::getResult() const
+{
+  CVC5_API_TRY_CATCH_BEGIN;
+  //////// all checks before this line
+  internal::NodeManager* nm = internal::NodeManager::currentNM();
+  return Term(nm, this->getProofNode()->getResult());
+  ////////
+  CVC5_API_TRY_CATCH_END;
+}
+
+const std::vector<Proof> Proof::getChildren() const
+{
+  CVC5_API_TRY_CATCH_BEGIN;
+  //////// all checks before this line
+  std::vector<Proof> children;
+  std::vector<std::shared_ptr<internal::ProofNode>> node_children =
+      d_proof_node->getChildren();
+  for (size_t i = 0, psize = node_children.size(); i < psize; i++)
+  {
+    children.push_back(Proof(node_children[i]));
+  }
+  return children;
+  ////////
+  CVC5_API_TRY_CATCH_END;
+}
+
+const std::vector<Term> Proof::getArguments() const
+{
+  CVC5_API_TRY_CATCH_BEGIN;
+  //////// all checks before this line
+  internal::NodeManager* nm = internal::NodeManager::currentNM();
+  std::vector<Term> args;
+  const std::vector<internal::Node> node_args = d_proof_node->getArguments();
+  for (size_t i = 0, asize = node_args.size(); i < asize; i++)
+  {
+    args.push_back(Term(nm, node_args[i]));
+  }
+  return args;
+  ////////
+  CVC5_API_TRY_CATCH_END;
+}
+
+const std::shared_ptr<internal::ProofNode>& Proof::getProofNode(void) const
+{
+  return this->d_proof_node;
+}
+
 /* -------------------------------------------------------------------------- */
 /* Solver                                                                     */
 /* -------------------------------------------------------------------------- */
@@ -7115,14 +7186,34 @@ std::pair<Result, std::vector<Term>> Solver::getTimeoutCore() const
   CVC5_API_TRY_CATCH_END;
 }
 
-std::string Solver::getProof(modes::ProofComponent c) const
+std::vector<Proof> Solver::getProof(modes::ProofComponent c) const
 {
   CVC5_API_TRY_CATCH_BEGIN;
   CVC5_API_CHECK(d_slv->getOptions().smt.produceProofs)
       << "Cannot get proof unless proofs are enabled (try --produce-proofs)";
   CVC5_API_RECOVERABLE_CHECK(d_slv->getSmtMode() == internal::SmtMode::UNSAT)
       << "Cannot get proof unless in unsat mode.";
-  return d_slv->getProof(c);
+  //////// all checks before this line
+  std::vector<std::shared_ptr<internal::ProofNode>> proof_nodes =
+      d_slv->getProof(c);
+  std::vector<Proof> proofs;
+  for (auto& p : proof_nodes)
+  {
+    proofs.push_back(Proof(p));
+  }
+  return proofs;
+  ////////
+  CVC5_API_TRY_CATCH_END;
+}
+
+std::string Solver::proofToString(Proof proof, modes::ProofFormat format) const
+{
+  CVC5_API_TRY_CATCH_BEGIN;
+  //////// all checks before this line
+  std::ostringstream ss;
+  this->d_slv->printProof(ss, proof.getProofNode(), format);
+  return ss.str();
+  ////////
   CVC5_API_TRY_CATCH_END;
 }